requirements.txt
paho-mqtt
influxdb
haversine
